wnsm-smartmeter/requirements.txt
requests>=2.26.0
lxml>=4.9.3
python-dateutil>=2.8.2
paho-mqtt>=1.6.1
PyYAML>=6.0
pytest>=7.0.0
git+https://github.com/cretl/vienna-smartmeter.git@fix-login-add-pkce